# packages/contract/openapi.yaml
openapi: 3.0.3
info:
  title: SmartKasi API
  version: 1.0.0-rc1
  description: |
    Hyperlocal spaza commerce for South African townships — offline-capable POS,
    a shared product catalog with cross-shop price comparison, multi-shop
    marketplace orders, and (v2) courier delivery.

    ## How to read this document

    **Every operation is tagged `[LIVE]` or `[STUB]`.**

    | Tag | Meaning | What clients may assume |
    |---|---|---|
    | `[LIVE]` | Real implementation, real database, ships Sat 22 Aug 10:00 SAST | Build against it. Behaviour is contractual. |
    | `[STUB]` | Route exists and returns a **fixed, schema-valid fake response**. No database, no logic. | Wire your UI to it. **Do not build business logic on the values.** Shapes will not change; values will. |

    Machine-readable: every operation also carries `x-smartkasi-status`.

    ## Frozen contract

    Once this file is handed to the client teams the **shapes are frozen**.
    Additive changes only (new optional fields, new endpoints). Any breaking
    change gets a new path version and an announcement in the team channel —
    not a silent edit.

    ## Non-negotiable rules

    1. **Money is integer cents (ZAR).** Field names always end `_cents`.
       `1850` is R18.50. Never send or expect a decimal.
    2. **Auth is Supabase.** This API issues no tokens and has no
       `/login`. See § Authentication.
    3. **Courier location and route are never returned to a customer.**
...

...

...

    ```
...
    ```

...

...
servers:
...
security:
  - supabaseJwt: []
...
paths:
...
  /shops:
    get:
      tags: [Shops]
      summary: "[LIVE] List / search shops near a point"
      description: |
        Powers the customer app's map and shop list. Results are ordered by
        distance ascending when `lat`/`lng` are supplied, otherwise by name.
      operationId: listShops
      x-smartkasi-status: LIVE
      security: []
      parameters:
        - { $ref: "#/components/parameters/Lat" }
        - { $ref: "#/components/parameters/Lng" }
        - { $ref: "#/components/parameters/RadiusM" }
        - name: q
          in: query
          description: Fuzzy match on shop name.
          schema: { type: string }
        - name: accepts_orders
          in: query
          description: Only shops that can take online orders (implies a verified licence).
          schema: { type: boolean }
        - name: open_now
          in: query
          schema: { type: boolean }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PerPage" }
      responses:
        "200":
          description: A page of shops.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Paginated"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/ShopSummary" }
              example:
                data:
                  - id: 7b0e1c2a-1111-4a3b-9c11-aaaaaaaaaaaa
                    name: Mama Thoko's Tuckshop
                    slug: mama-thokos-tuckshop
                    logo_url: https://cdn.smartkasi.co.za/shops/thoko.jpg
                    address_line: 1423 Vilakazi St
                    township: Orlando West
                    city: Soweto
                    lat: -26.2380
                    lng: 27.9083
                    distance_m: 480
                    mode: full
                    accepts_orders: true
                    accepts_delivery: true
                    is_open_now: true
                  - id: 7b0e1c2a-2222-4a3b-9c11-bbbbbbbbbbbb
                    name: Bra Sipho Spaza
                    slug: bra-sipho-spaza
                    logo_url: null
                    address_line: 88 Mooki St
                    township: Orlando East
                    city: Soweto
                    lat: -26.2461
                    lng: 27.9212
                    distance_m: 1620
                    mode: advertising_only
                    accepts_orders: false
                    accepts_delivery: false
                    is_open_now: true
                meta: { page: 1, per_page: 25, total: 2, total_pages: 1 }
    post:
      tags: [Shops]
      summary: "[LIVE] Register a shop"
      description: |
        Creates the shop in `advertising_only` mode. `accepts_orders` cannot be
        set here — it unlocks only after the trading licence is verified via
        `POST /shops/{shopId}/licence`.
      operationId: createShop
      x-smartkasi-status: LIVE
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateShopRequest" }
      responses:
        "201":
          description: Shop created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shop" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409":
          description: A shop with that slug already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /shops/{shopId}:
    parameters:
      - { $ref: "#/components/parameters/ShopId" }
    get:
      tags: [Shops]
      summary: "[LIVE] Get a shop"
      operationId: getShop
      x-smartkasi-status: LIVE
      security: []
      responses:
        "200":
          description: Shop detail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shop" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Shops]
      summary: "[LIVE] Update a shop"
      description: "Owner only. `accepts_orders: true` is rejected unless `licence_status = verified`."
      operationId: updateShop
      x-smartkasi-status: LIVE
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateShopRequest" }
      responses:
        "200":
          description: Updated shop.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shop" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: "`LICENCE_NOT_VERIFIED` — cannot enable ordering yet."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /shops/{shopId}/licence:
    parameters:
      - { $ref: "#/components/parameters/ShopId" }
    post:
      tags: [Shops]
      summary: "[LIVE] Submit trading licence for verification"
      description: |
        Upload the document to R2 first via `POST /uploads/presign`, then submit
        the resulting URL here. Sets `licence_status` to `pending`. In v1 an
        admin verifies manually from the dashboard.
      operationId: submitLicence
      x-smartkasi-status: LIVE
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [trading_licence_no, licence_doc_url]
              properties:
                trading_licence_no: { type: string, example: "GP/SOW/2026/00841" }
                licence_doc_url: { type: string, format: uri }
                licence_expires_at: { type: string, format: date }
      responses:
        "202":
          description: Submitted, awaiting review.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Shop" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ===========================================================================
  # CATALOG
  # ===========================================================================
...
  /shops/{shopId}/sales:
    parameters:
      - { $ref: "#/components/parameters/ShopId" }
    get:
      tags: [Sales]
      summary: "[LIVE] List sales"
      operationId: listSales
      x-smartkasi-status: LIVE
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: cashier_id
          in: query
          schema: { type: string, format: uuid }
        - { $ref: "#/components/parameters/Page" }
        - { $ref: "#/components/parameters/PerPage" }
      responses:
        "200":
          description: A page of sales.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Paginated"
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/Sale" }
    post:
      tags: [Sales]
      summary: "[LIVE] Record a single sale (online till)"
      description: |
        Idempotent on `client_sale_id`. Reposting the same `client_sale_id`
        returns `200` with the original sale rather than creating a second one.
      operationId: createSale
      x-smartkasi-status: LIVE
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSaleRequest" }
      responses:
        "200":
          description: Already recorded; returning the original.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sale" }
        "201":
          description: Sale recorded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sale" }
        "422":
          description: "`INSUFFICIENT_STOCK` or `TOTALS_MISMATCH`."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /shops/{shopId}/sales/batch:
    parameters:
      - { $ref: "#/components/parameters/ShopId" }
    post:
      tags: [Sales, Sync]
      summary: "[LIVE] Flush queued offline sales"
      description: |
        The offline workhorse. Max 200 sales per call.

        **Partial success is the normal case.** Always read `results` item by
        item; never treat `207` as failure. Rows returned as `duplicate` are a
        success — mark them synced locally and stop resending.

        Only clear a sale from the device queue when its `status` is `created`
        or `duplicate`. `failed` rows stay queued for human review.
      operationId: batchCreateSales
      x-smartkasi-status: LIVE
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sales]
              properties:
                sales:
                  type: array
                  maxItems: 200
                  items: { $ref: "#/components/schemas/CreateSaleRequest" }
            example:
              sales:
                - client_sale_id: 11111111-1111-4111-8111-111111111111
                  sold_at: "2026-08-21T14:32:11Z"
                  payment_method: cash
                  subtotal_cents: 3450
                  discount_cents: 0
                  total_cents: 3450
                  amount_tendered_cents: 5000
                  change_cents: 1550
                  items:
                    - { product_id: 3f0a9d10-aaaa-4c11-9999-111111111111, qty: 1, unit_price_cents: 2000 }
                    - { product_id: 3f0a9d10-bbbb-4c11-9999-222222222222, qty: 2, unit_price_cents: 725 }
      responses:
        "207":
          description: Per-sale outcome.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BatchSaleResponse" }
              example:
                server_time: "2026-08-22T09:14:02Z"
                summary: { created: 1, duplicate: 1, failed: 1 }
                results:
                  - client_sale_id: 11111111-1111-4111-8111-111111111111
                    status: created
                    sale_id: 9c8b7a60-0000-4000-8000-000000000001
                  - client_sale_id: 22222222-2222-4222-8222-222222222222
                    status: duplicate
                    sale_id: 9c8b7a60-0000-4000-8000-000000000002
                  - client_sale_id: 33333333-3333-4333-8333-333333333333
                    status: failed
                    sale_id: null
                    error:
                      code: TOTALS_MISMATCH
                      message: "total_cents 3400 does not equal subtotal 3450 minus discount 0"
  /shops/{shopId}/sales/{saleId}:
    parameters:
      - { $ref: "#/components/parameters/ShopId" }
      - name: saleId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [Sales]
      summary: "[LIVE] Get a sale (receipt)"
      operationId: getSale
      x-smartkasi-status: LIVE
      responses:
        "200":
          description: Sale with line items.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sale" }
        "404": { $ref: "#/components/responses/NotFound" }
  /shops/{shopId}/sales/{saleId}/void:
    parameters:
      - { $ref: "#/components/parameters/ShopId" }
      - name: saleId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      tags: [Sales]
      summary: "[LIVE] Void a sale"
      description: |
        Marks the sale `voided` and writes compensating `void` movements to the
        stock ledger. The sale row is never deleted — voids must be auditable.
        Requires `can_void_sales`.
      operationId: voidSale
      x-smartkasi-status: LIVE
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, maxLength: 240 }
      responses:
        "200":
          description: Voided.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sale" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /shops/{shopId}/reports/daily:
    parameters:
      - { $ref: "#/components/parameters/ShopId" }
    get:
      tags: [Sales]
      summary: "[LIVE] Daily takings summary"
      description: The cash-up screen. Defaults to today in Africa/Johannesburg.
      operationId: getDailyReport
      x-smartkasi-status: LIVE
      parameters:
        - name: date
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Summary for the day.
          content:
            application/json:
              schema:
                type: object
                properties:
                  date: { type: string, format: date }
                  sale_count: { type: integer }
                  gross_cents: { type: integer, format: int64 }
                  discount_cents: { type: integer, format: int64 }
                  net_cents: { type: integer, format: int64 }
                  voided_count: { type: integer }
                  by_payment_method:
                    type: array
                    items:
                      type: object
                      properties:
                        method: { type: string, enum: [cash, card, qr, account] }
                        count: { type: integer }
                        total_cents: { type: integer, format: int64 }
                  top_products:
                    type: array
                    items:
                      type: object
                      properties:
                        product_id: { type: string, format: uuid }
                        name: { type: string }
                        qty: { type: integer }
                        total_cents: { type: integer, format: int64 }
              example:
                date: "2026-08-21"
                sale_count: 87
                gross_cents: 412350
                discount_cents: 0
                net_cents: 412350
                voided_count: 2
                by_payment_method:
                  - { method: cash, count: 81, total_cents: 388900 }
                  - { method: card, count: 6, total_cents: 23450 }
                top_products:
                  - { product_id: 3f0a9d10-aaaa-4c11-9999-111111111111, name: Iwisa Super Maize Meal, qty: 12, total_cents: 102000 }

  # ===========================================================================
  # ORDERS
  # ===========================================================================
...
  /shops/{shopId}/flyers:
    parameters:
      - { $ref: "#/components/parameters/ShopId" }
    get:
      tags: [Flyers]
      summary: "[LIVE] List a shop's active flyers"
      operationId: listFlyers
      x-smartkasi-status: LIVE
      security: []
      parameters:
        - name: active_only
          in: query
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: Flyers.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Flyer" }
    post:
      tags: [Flyers]
      summary: "[LIVE] Upload a flyer"
      description: |
        The entire value proposition for an `advertising_only` shop — no POS, no
        inventory, just a photo of this week's specials. Upload the image via
        `POST /uploads/presign` first.
      operationId: createFlyer
      x-smartkasi-status: LIVE
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, image_url, starts_at, ends_at]
              properties:
                title: { type: string, maxLength: 120 }
                image_url: { type: string, format: uri }
                starts_at: { type: string, format: date }
                ends_at: { type: string, format: date }
      responses:
        "201":
          description: Flyer created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Flyer" }
  /shops/{shopId}/flyers/{flyerId}:
    parameters:
      - { $ref: "#/components/parameters/ShopId" }
      - name: flyerId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [Flyers]
      summary: "[LIVE] Remove a flyer"
      operationId: deleteFlyer
      x-smartkasi-status: LIVE
      responses:
        "204": { description: Deleted. }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ===========================================================================
  # UPLOADS
  # ===========================================================================
...
components:
  securitySchemes:
    supabaseJwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase access token. Obtained client-side via the Supabase SDK; this API never issues one.
  parameters:
    ShopId:
      name: shopId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Lat:
      name: lat
      in: query
      description: Latitude of the search origin. Required together with `lng` for distance sorting.
      schema: { type: number, format: double, example: -26.2380 }
    Lng:
      name: lng
      in: query
      schema: { type: number, format: double, example: 27.9083 }
    RadiusM:
      name: radius_m
      in: query
      description: Search radius in metres. Capped at 10000.
      schema: { type: integer, minimum: 100, maximum: 10000, default: 2000 }
  responses:
    BadRequest:
      description: Malformed request. `error.code` is `VALIDATION_FAILED`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Missing, malformed or expired token. `error.code` is `UNAUTHENTICATED` or `TOKEN_EXPIRED`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: Authenticated but not permitted. `error.code` is `FORBIDDEN`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    ErrorDetail:
      type: object
      properties:
        field: { type: string }
        issue: { type: string }
        code: { type: string }
        message: { type: string }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: |
                Stable machine-readable code. Switch on this, never on `message`.
              enum:
                - VALIDATION_FAILED
                - UNAUTHENTICATED
                - TOKEN_EXPIRED
                - FORBIDDEN
                - NOT_FOUND
                - PRODUCT_NOT_FOUND
                - ALREADY_STOCKED
                - STALE_WRITE
                - INSUFFICIENT_STOCK
                - TOTALS_MISMATCH
                - DUPLICATE_SALE
                - LICENCE_NOT_VERIFIED
                - SHOP_NOT_ACCEPTING_ORDERS
                - QUOTE_EXPIRED
                - SHOPS_TOO_FAR_APART
                - OUTSIDE_SERVICE_AREA
                - ORDER_NOT_CANCELLABLE
                - RATE_LIMITED
                - NOT_IMPLEMENTED
                - INTERNAL_ERROR
            message: { type: string }
            details:
              type: array
              items: { $ref: "#/components/schemas/ErrorDetail" }
            request_id:
              type: string
              description: Include this when reporting a bug — it finds the log line.
    Paginated:
      type: object
      required: [data, meta]
      properties:
        data: { type: array, items: {} }
        meta:
          type: object
          required: [page, per_page, total, total_pages]
          properties:
            page: { type: integer }
            per_page: { type: integer }
            total: { type: integer }
            total_pages: { type: integer }

    # ---- Enums ------------------------------------------------------------
...
    LicenceStatus:
      type: string
      enum: [none, pending, verified, rejected, expired]
    ShopMode:
      type: string
      enum: [advertising_only, inventory_only, full]
    PaymentMethod:
      type: string
      enum: [cash, card, qr, account]
...
    ShopSummary:
      type: object
      required: [id, name, slug, address_line, lat, lng, mode, accepts_orders]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        logo_url: { type: string, nullable: true }
        address_line: { type: string }
        township: { type: string, nullable: true }
        city: { type: string, nullable: true }
        lat: { type: number, format: double }
        lng: { type: number, format: double }
        distance_m:
          type: integer
          nullable: true
          description: Present only when the request supplied `lat`/`lng`.
        mode: { $ref: "#/components/schemas/ShopMode" }
        accepts_orders: { type: boolean }
        accepts_delivery: { type: boolean }
        is_open_now: { type: boolean }
    Shop:
      allOf:
        - $ref: "#/components/schemas/ShopSummary"
        - type: object
          properties:
            owner_id: { type: string, format: uuid }
            description: { type: string, nullable: true }
            phone: { type: string, nullable: true }
            province: { type: string, nullable: true }
            trading_licence_no: { type: string, nullable: true }
            licence_status: { $ref: "#/components/schemas/LicenceStatus" }
            licence_expires_at: { type: string, format: date, nullable: true }
            opens_at: { type: string, nullable: true, example: "07:00" }
            closes_at: { type: string, nullable: true, example: "20:00" }
            is_active: { type: boolean }
            product_count: { type: integer }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    CreateShopRequest:
      type: object
      required: [name, address_line, lat, lng]
      properties:
        name: { type: string, maxLength: 120, example: Mama Thoko's Tuckshop }
        description: { type: string }
        phone: { type: string, example: "+27821234567" }
        address_line: { type: string, example: 1423 Vilakazi St }
        township: { type: string, example: Orlando West }
        city: { type: string, example: Soweto }
        province: { type: string, example: Gauteng }
        lat: { type: number, format: double, example: -26.2380 }
        lng: { type: number, format: double, example: 27.9083 }
        opens_at: { type: string, example: "07:00" }
        closes_at: { type: string, example: "20:00" }
    UpdateShopRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        phone: { type: string }
        logo_url: { type: string, format: uri }
        address_line: { type: string }
        lat: { type: number, format: double }
        lng: { type: number, format: double }
        opens_at: { type: string }
        closes_at: { type: string }
        mode: { $ref: "#/components/schemas/ShopMode" }
        accepts_orders:
          type: boolean
          description: Rejected with `LICENCE_NOT_VERIFIED` unless the trading licence is verified.
        accepts_delivery: { type: boolean }
        is_active: { type: boolean }

    # ---- Catalog ----------------------------------------------------------
...
    CreateSaleRequest:
      type: object
      required: [client_sale_id, sold_at, subtotal_cents, total_cents, items]
      properties:
        client_sale_id:
          type: string
          format: uuid
          description: Generated on the device. The idempotency key. Never reuse, never regenerate on retry.
        sold_at: { type: string, format: date-time, description: Client clock. }
        payment_method: { $ref: "#/components/schemas/PaymentMethod" }
        subtotal_cents: { type: integer, format: int64, minimum: 0 }
        discount_cents: { type: integer, format: int64, minimum: 0, default: 0 }
        total_cents:
          type: integer
          format: int64
          minimum: 0
          description: Must equal `subtotal_cents - discount_cents`, else `TOTALS_MISMATCH`.
        amount_tendered_cents: { type: integer, format: int64, minimum: 0 }
        change_cents: { type: integer, format: int64, minimum: 0 }
        items:
          type: array
          minItems: 1
          items: { $ref: "#/components/schemas/SaleItemInput" }
...
    Sale:
      type: object
      properties:
        id: { type: string, format: uuid }
        shop_id: { type: string, format: uuid }
        client_sale_id: { type: string, format: uuid }
        cashier_id: { type: string, format: uuid, nullable: true }
        cashier_name: { type: string, nullable: true }
        status: { type: string, enum: [completed, voided] }
        subtotal_cents: { type: integer, format: int64 }
        discount_cents: { type: integer, format: int64 }
        total_cents: { type: integer, format: int64 }
        amount_tendered_cents: { type: integer, format: int64, nullable: true }
        change_cents: { type: integer, format: int64, nullable: true }
        payment_method: { $ref: "#/components/schemas/PaymentMethod" }
        items:
          type: array
          items: { $ref: "#/components/schemas/SaleItem" }
        sold_at: { type: string, format: date-time }
        synced_at: { type: string, format: date-time }
      example:
        id: 9c8b7a60-0000-4000-8000-000000000001
        shop_id: 7b0e1c2a-1111-4a3b-9c11-aaaaaaaaaaaa
        client_sale_id: 11111111-1111-4111-8111-111111111111
        cashier_id: 11111111-0000-4000-8000-000000000001
        cashier_name: Thoko Ndlovu
        status: completed
        subtotal_cents: 3450
        discount_cents: 0
        total_cents: 3450
        amount_tendered_cents: 5000
        change_cents: 1550
        payment_method: cash
        items:
          - id: aa000000-0000-4000-8000-000000000001
            product_id: 3f0a9d10-aaaa-4c11-9999-111111111111
            product_name: Iwisa Super Maize Meal
            qty: 1
            unit_price_cents: 2000
            line_total_cents: 2000
          - id: aa000000-0000-4000-8000-000000000002
            product_id: 3f0a9d10-bbbb-4c11-9999-222222222222
            product_name: Sunlight Bar Soap 250g
            qty: 2
            unit_price_cents: 725
            line_total_cents: 1450
        sold_at: "2026-08-21T14:32:11Z"
        synced_at: "2026-08-22T09:14:02Z"
    BatchSaleResponse:
      type: object
      required: [server_time, summary, results]
      properties:
        server_time: { type: string, format: date-time }
        summary:
          type: object
          properties:
            created: { type: integer }
            duplicate: { type: integer }
            failed: { type: integer }
        results:
          type: array
          items:
            type: object
            required: [client_sale_id, status]
            properties:
              client_sale_id: { type: string, format: uuid }
              status:
                type: string
                enum: [created, duplicate, failed]
                description: "`duplicate` is a SUCCESS. Clear it from the device queue."
              sale_id: { type: string, format: uuid, nullable: true }
              error: { $ref: "#/components/schemas/ErrorDetail" }

    # ---- Orders -----------------------------------------------------------
...
    Flyer:
      type: object
      properties:
        id: { type: string, format: uuid }
        shop_id: { type: string, format: uuid }
        title: { type: string }
        image_url: { type: string, format: uri }
        starts_at: { type: string, format: date }
        ends_at: { type: string, format: date }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }

    # ---- AI ---------------------------------------------------------------
...